% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrino.service.R
\name{constructURL}
\alias{constructURL}
\title{construct the url for the api get request}
\usage{
constructURL(endpoint, next_page = NULL)
}
\description{
construct the url for the api get request
}
\keyword{internal}
