% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company.service.R
\name{getAllCompanies}
\alias{getAllCompanies}
\title{All Companies
Returns all Companies. When parameters are specified, returns matching companies.}
\usage{
getAllCompanies(opts = list())
}
\arguments{
\item{opts}{{named list} Optional parameters}

\item{opts$latestFilingDate}{{String}  YYYY-MM-DD Last filing date}

\item{opts$sic}{{String}  Return companies with the given Standard Industrial Classification code}

\item{opts$template}{{module:model/String}  Return companies with the given financial statement template}

\item{opts$sector}{{String}  Return companies in the given industry sector}

\item{opts$industryCategory}{{String}  Return companies in the given industry category}

\item{opts$industryGroup}{{String}  Return companies in the given industry group}

\item{opts$hasFundamentals}{{Boolean}  Return only companies that have fundamentals when true}

\item{opts$hasStockPrices}{{Boolean}  Return only companies that have stock prices when true}

\item{opts$pageSize}{{Number} The number of results to return (default to 100)}
}
\value{
companies {tibble}
}
\description{
All Companies
Returns all Companies. When parameters are specified, returns matching companies.
}
\seealso{
Other company endpoints: \code{\link{getCompanyDataPointHistory}},
  \code{\link{getCompanyDataPointNumber}},
  \code{\link{getCompanyDataPointText}},
  \code{\link{getCompanyFilings}},
  \code{\link{getCompanyFundamentals}},
  \code{\link{getCompanyIPOs}},
  \code{\link{getCompanyNews}},
  \code{\link{getCompanySecurities}},
  \code{\link{getCompany}}
}
\concept{company endpoints}
