% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicies.R
\name{getAllEconomicIndicies}
\alias{getAllEconomicIndicies}
\title{Return Econpmic Indices Meta-Data}
\usage{
getAllEconomicIndicies(opts = list())
}
\arguments{
\item{opts}{{named list} Optional query parameters}

\item{opts$page_size}{the number of results to return (defaults to 100)}
}
\description{
Return Econpmic Indices Meta-Data
}
\seealso{
Other indices endpoints: \code{\link{getAllSicIndicies}},
  \code{\link{getAllStockMarketIndicies}},
  \code{\link{getIndicesDataPointNumber}},
  \code{\link{getIndicesDataPointText}}
}
\concept{indices endpoints}
