% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicies.R
\name{getAllSicIndicies}
\alias{getAllSicIndicies}
\title{Return SIC Indices Meta-Data}
\usage{
getAllSicIndicies(opts = list())
}
\arguments{
\item{opts}{{named list} Optional query parameters}

\item{opts$page_size}{the number of results to return (defaults to 100)}
}
\description{
Return SIC Indices Meta-Data
}
\seealso{
Other indices endpoints: \code{\link{getAllEconomicIndicies}},
  \code{\link{getAllStockMarketIndicies}},
  \code{\link{getIndicesDataPointNumber}},
  \code{\link{getIndicesDataPointText}}
}
\concept{indices endpoints}
