% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company.service.R
\name{getCompanyById}
\alias{getCompanyById}
\title{Returns the Company with the given \code{identifier}}
\usage{
getCompanyById(identifier)
}
\arguments{
\item{identifier}{A Company identifier (Ticker, CIK, LEI, Intrinio ID)}
}
\value{
df {tibble}
}
\description{
Returns the Company with the given \code{identifier}
}
\seealso{
Other security endpoints: \code{\link{getAllSecurities}},
  \code{\link{getSecurityById}},
  \code{\link{getSecurityDataPointHistory}},
  \code{\link{getSecurityDataPointNumber}},
  \code{\link{getSecurityDataPointText}},
  \code{\link{getSecurityDividendsLatest}},
  \code{\link{getSecurityEarningsLatest}},
  \code{\link{getSecurityRealtimePrice}},
  \code{\link{getSecurityStockPrices}},
  \code{\link{getSecurityZacksAnalystRatingsSnapshot}}
}
\concept{security endpoints}
