% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company.service.R
\name{getCompanyDataPointHistory}
\alias{getCompanyDataPointHistory}
\title{historical values data-point tag for a company over specified time range}
\usage{
getCompanyDataPointHistory(identifier, tag, opts = list())
}
\arguments{
\item{identifier}{{String} A company identifier (Ticker, FIGI, ISIN, CUSIP, Intrinio ID)}

\item{tag}{{String} An Intrinio data tag ID or code}

\item{opts}{{named list} Optional query params for endpoint}

\item{opts$start_date}{{Date} YYYY-MM-DD Return prices on or after the date}

\item{opts$end_date':}{{Date} YYYY-MM-DD Return prices on or before the date}

\item{opts$sort_order}{{String = "desc"} desc | asc}

\item{opts$frequency:}{{String = "daily"} daily | weekly | monthly | quarterly | yearly Return stock prices in the given frequency}

\item{opts$type}{{String} FY | QTR | TTM Return historical data for given fiscal period type}

\item{opts$page_size':}{{number = 100} The number of results to return per page}
}
\description{
historical values data-point tag for a company over specified time range
}
\seealso{
\url{https://data.intrinio.com/data-tags}

Other company endpoints: \code{\link{getAllCompanies}},
  \code{\link{getCompanyDataPointNumber}},
  \code{\link{getCompanyDataPointText}},
  \code{\link{getCompanyIPOs}},
  \code{\link{getCompanyNews}},
  \code{\link{getCompanySecurities}}
}
\concept{company endpoints}
