% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company.service.R
\name{getCompanyDataPointNumber}
\alias{getCompanyDataPointNumber}
\title{current numerical value data-point tag for a company}
\usage{
getCompanyDataPointNumber(identifier, tag)
}
\arguments{
\item{identifier}{{String} A company identifier (Ticker, FIGI, ISIN, CUSIP, Intrinio ID)}

\item{tag}{{String} An Intrinio data tag ID or code}
}
\value{
{numeric}
}
\description{
current numerical value data-point tag for a company
}
\seealso{
\url{https://data.intrinio.com/data-tags}

Other company endpoints: \code{\link{getAllCompanies}},
  \code{\link{getCompanyDataPointHistory}},
  \code{\link{getCompanyDataPointText}},
  \code{\link{getCompanyIPOs}},
  \code{\link{getCompanyNews}},
  \code{\link{getCompanySecurities}}
}
\concept{company endpoints}
