% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company.service.R
\name{getCompanyDataPointText}
\alias{getCompanyDataPointText}
\title{current text value of data-point tag for a company}
\usage{
getCompanyDataPointText(identifier, tag)
}
\arguments{
\item{identifier}{{String} A company identifier (Ticker, FIGI, ISIN, CUSIP, Intrinio ID)}

\item{tag}{{String} An Intrinio data tag ID or code}
}
\value{
{String}
}
\description{
current text value of data-point tag for a company
}
\seealso{
\url{https://data.intrinio.com/data-tags}

Other company endpoints: \code{\link{getAllCompanies}},
  \code{\link{getCompanyDataPointHistory}},
  \code{\link{getCompanyDataPointNumber}},
  \code{\link{getCompanyFilings}},
  \code{\link{getCompanyFundamentals}},
  \code{\link{getCompanyIPOs}},
  \code{\link{getCompanyNews}},
  \code{\link{getCompanySecurities}},
  \code{\link{getCompany}}

Other company endpoints: \code{\link{getAllCompanies}},
  \code{\link{getCompanyDataPointHistory}},
  \code{\link{getCompanyDataPointNumber}},
  \code{\link{getCompanyFilings}},
  \code{\link{getCompanyFundamentals}},
  \code{\link{getCompanyIPOs}},
  \code{\link{getCompanyNews}},
  \code{\link{getCompanySecurities}},
  \code{\link{getCompany}}
}
\concept{company endpoints}
