% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company.service.R
\name{getCompanyNews}
\alias{getCompanyNews}
\title{Returns news for the Company with the given \code{identifier}}
\usage{
getCompanyNews(identifier, opts = list(), max_n = 10)
}
\arguments{
\item{identifier}{A Company identifier (Ticker, CIK, LEI, Intrinio ID)}

\item{opts}{{named list} Optional query parameters}

\item{max_n}{{number} maximum results to return (defaults to 10)}

\item{opts$page_size}{{number} Number of results returned per page default = 100}
}
\value{
companies {tibble}
}
\description{
Returns news for the Company with the given \code{identifier}
}
\seealso{
Other company endpoints: \code{\link{getAllCompanies}},
  \code{\link{getCompanyDataPointHistory}},
  \code{\link{getCompanyDataPointNumber}},
  \code{\link{getCompanyDataPointText}},
  \code{\link{getCompanyIPOs}},
  \code{\link{getCompanySecurities}}
}
\concept{company endpoints}
