% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company.service.R
\name{getCompanySecurities}
\alias{getCompanySecurities}
\title{Returns Securities for the Company with the given \code{identifier}}
\usage{
getCompanySecurities(identifier)
}
\arguments{
\item{identifier}{A Company identifier (Ticker, CIK, LEI, Intrinio ID)}
}
\value{
companies {tibble}
}
\description{
Returns Securities for the Company with the given \code{identifier}
}
\seealso{
Other company endpoints: \code{\link{getAllCompanies}},
  \code{\link{getCompanyDataPointHistory}},
  \code{\link{getCompanyDataPointNumber}},
  \code{\link{getCompanyDataPointText}},
  \code{\link{getCompanyFilings}},
  \code{\link{getCompanyFundamentals}},
  \code{\link{getCompanyIPOs}},
  \code{\link{getCompanyNews}}, \code{\link{getCompany}}
}
\concept{company endpoints}
