% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filings.R
\name{getFilingById}
\alias{getFilingById}
\title{Returns the Filing meta data for the given \code{id}}
\usage{
getFilingById(id)
}
\arguments{
\item{id}{{String}  The Intrinio ID for the Filing}
}
\description{
Returns the Filing meta data for the given \code{id}
}
\concept{filing endpoints}
