% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.service.R
\name{getFundamentalById}
\alias{getFundamentalById}
\title{Returns detailed fundamental financial data for the given \code{id}.}
\usage{
getFundamentalById(id)
}
\arguments{
\item{id}{{String}  The Intrinio ID for the Fundamental}
}
\description{
Returns detailed fundamental financial data for the given \code{id}.
}
\seealso{
Other fundamental endpoints: \code{\link{getReportedFinancialsTable}},
  \code{\link{getReportedFinancials}},
  \code{\link{getStandardizedFinancialsTable}},
  \code{\link{getStandardizedFinancials}},
  \code{\link{lookupFundamental}}
}
\concept{fundamental endpoints}
