% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicies.R
\name{getIndicesDataPointHistory}
\alias{getIndicesDataPointHistory}
\title{historical values data-point tag for a index over specified time range}
\usage{
getIndicesDataPointHistory(identifier, tag = "level",
  type = "stock_market", opts = list(), rm.na = TRUE)
}
\arguments{
\item{identifier}{{String} A Index identifier (symbol, Intrinio ID)}

\item{tag}{{String} An Intrinio data tag ID or code (defaults to `level``)}

\item{opts}{{named list} Optional query params for endpoint}

\item{opts$start_date}{{Date} YYYY-MM-DD Return prices on or after the date}

\item{opts$end_date':}{{Date} YYYY-MM-DD Return prices on or before the date}

\item{opts$sort_order}{{String = "desc"} desc | asc}

\item{opts$page_size':}{{number = 100} The number of results to return per page}
}
\description{
Stock Market Indices: "$SPX"   "$DJI"   "$DJA"   "$TRAN"  "$UTIL"  "$NDX"   "$COMPQ"
}
\seealso{
\url{https://data.intrinio.com/data-tags}

Other security endpoints: \code{\link{getAllSecurities}},
  \code{\link{getCompany}}, \code{\link{getSecurityById}},
  \code{\link{getSecurityDataPointHistory}},
  \code{\link{getSecurityDataPointNumber}},
  \code{\link{getSecurityDataPointText}},
  \code{\link{getSecurityDividendsLatest}},
  \code{\link{getSecurityEarningsLatest}},
  \code{\link{getSecurityIntradayPrices}},
  \code{\link{getSecurityRealtimePrice}},
  \code{\link{getSecurityStockPrices}},
  \code{\link{getSecurityZacksAnalystRatingsSnapshot}}
}
\concept{security endpoints}
