% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicies.R
\name{getIndicesDataPointNumber}
\alias{getIndicesDataPointNumber}
\title{current numerical value data-point tag for an index}
\usage{
getIndicesDataPointNumber(identifier, tag = "level",
  type = "stock_market")
}
\arguments{
\item{identifier}{{String} A Index identifier (symbol, Intrinio ID)}

\item{tag}{{String} An Intrinio data tag ID or code}
}
\value{
{numeric}
}
\description{
current numerical value data-point tag for an index
}
\seealso{
\url{https://data.intrinio.com/data-tags}

Other indices endpoints: \code{\link{getAllEconomicIndicies}},
  \code{\link{getAllSicIndicies}},
  \code{\link{getAllStockMarketIndicies}},
  \code{\link{getIndicesDataPointText}}
}
\concept{indices endpoints}
