% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicies.R
\name{getIndicesDataPointText}
\alias{getIndicesDataPointText}
\title{current text value of data-point tag for an index}
\usage{
getIndicesDataPointText(identifier, tag = "level",
  index = "stock_market")
}
\arguments{
\item{identifier}{{String} A Security identifier (index symbol, Intrinio ID)}

\item{tag}{{String} An Intrinio data tag ID or code}
}
\value{
{String}
}
\description{
current text value of data-point tag for an index
}
\seealso{
\url{https://data.intrinio.com/data-tags}

Other indices endpoints: \code{\link{getAllEconomicIndicies}},
  \code{\link{getAllSicIndicies}},
  \code{\link{getAllStockMarketIndicies}},
  \code{\link{getIndicesDataPointNumber}}
}
\concept{indices endpoints}
