% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.service.R
\name{getReportedFinancialsTable}
\alias{getReportedFinancialsTable}
\title{Returns the As-Reported Financials directly from the financial
statements of the XBRL filings from the company specified by the \code{id} param.}
\usage{
getReportedFinancialsTable(id)
}
\arguments{
\item{id}{{String}  The Intrinio ID for the Fundamental or lookup code (ticker-statement-year-period)
for the Fundamental}
}
\value{
reported_financials {tibble} formated as wide table with one column per metric tag
}
\description{
Returns the As-Reported Financials directly from the financial
statements of the XBRL filings from the company specified by the \code{id} param.
}
\seealso{
Other fundamental endpoints: \code{\link{getFundamentalById}},
  \code{\link{getReportedFinancials}},
  \code{\link{getStandardizedFinancialsTable}},
  \code{\link{getStandardizedFinancials}},
  \code{\link{lookupFundamental}}
}
\concept{fundamental endpoints}
