% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.service.R
\name{getSecurityDataPointHistory}
\alias{getSecurityDataPointHistory}
\title{Returns time-series of historical values data-point tag for a security over specified time range}
\usage{
getSecurityDataPointHistory(identifier, tag, opts = list())
}
\arguments{
\item{identifier}{{String} A Security identifier (Ticker, FIGI, ISIN, CUSIP, Intrinio ID)}

\item{tag}{{String} An Intrinio data tag ID or code}

\item{opts}{{named list} Optional query params for endpoint}

\item{opts$start_date}{{Date} YYYY-MM-DD Return prices on or after the date}

\item{opts$end_date}{{Date} YYYY-MM-DD Return prices on or before the date}

\item{opts$sort_order}{{String = "desc"} desc | asc}

\item{opts$frequency}{{String = "daily"} daily | weekly | monthly | quarterly | yearly Return stock prices in the given frequency}

\item{opts$page_size}{{number = 100} The number of results to return per page}
}
\description{
Returns time-series of historical values data-point tag for a security over specified time range
}
\seealso{
\url{https://data.intrinio.com/data-tags}

Other security endpoints: \code{\link{getAllSecurities}},
  \code{\link{getIndicesDataPointHistory}},
  \code{\link{getSecurityById}},
  \code{\link{getSecurityDataPointNumber}},
  \code{\link{getSecurityDataPointText}},
  \code{\link{getSecurityDividendsLatest}},
  \code{\link{getSecurityEarningsLatest}},
  \code{\link{getSecurityIntradayPrices}},
  \code{\link{getSecurityRealtimePrice}},
  \code{\link{getSecurityStockPrices}},
  \code{\link{getSecurityZacksAnalystRatingsSnapshot}}
}
\concept{security endpoints}
