% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.service.R
\name{getSecurityDataPointText}
\alias{getSecurityDataPointText}
\title{current text value of data-point tag for a security}
\usage{
getSecurityDataPointText(identifier, tag)
}
\arguments{
\item{identifier}{{String} A Security identifier (Ticker, FIGI, ISIN, CUSIP, Intrinio ID)}

\item{tag}{{String} An Intrinio data tag ID or code}
}
\value{
{String}
}
\description{
current text value of data-point tag for a security
}
\seealso{
\url{https://data.intrinio.com/data-tags}

Other security endpoints: \code{\link{getAllSecurities}},
  \code{\link{getIndicesDataPointHistory}},
  \code{\link{getSecurityById}},
  \code{\link{getSecurityDataPointHistory}},
  \code{\link{getSecurityDataPointNumber}},
  \code{\link{getSecurityDividendsLatest}},
  \code{\link{getSecurityEarningsLatest}},
  \code{\link{getSecurityIntradayPrices}},
  \code{\link{getSecurityRealtimePrice}},
  \code{\link{getSecurityStockPrices}},
  \code{\link{getSecurityZacksAnalystRatingsSnapshot}}
}
\concept{security endpoints}
