% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.service.R
\name{getSecurityDividendsLatest}
\alias{getSecurityDividendsLatest}
\title{Returns the latest available dividend information for the Security with the given \code{identifier}}
\usage{
getSecurityDividendsLatest(identifier)
}
\arguments{
\item{identifier}{{String} A Security identifier (Ticker, FIGI, ISIN, CUSIP, Intrinio ID)}
}
\description{
Returns the latest available dividend information for the Security with the given \code{identifier}
}
\seealso{
Other security endpoints: \code{\link{getAllSecurities}},
  \code{\link{getSecurityById}},
  \code{\link{getSecurityDataPointHistory}},
  \code{\link{getSecurityDataPointNumber}},
  \code{\link{getSecurityDataPointText}},
  \code{\link{getSecurityEarningsLatest}},
  \code{\link{getSecurityRealtimePrice}},
  \code{\link{getSecurityStockPrices}},
  \code{\link{getSecurityZacksAnalystRatingsSnapshot}}
}
\concept{security endpoints}
