% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.service.R
\name{getSecurityEarningsLatest}
\alias{getSecurityEarningsLatest}
\title{Returns latest available earnings information for the Security with the given \code{identifier}}
\usage{
getSecurityEarningsLatest(identifier)
}
\arguments{
\item{identifier}{{String} A Security identifier (Ticker, FIGI, ISIN, CUSIP, Intrinio ID)}
}
\description{
Returns latest available earnings information for the Security with the given \code{identifier}
}
\seealso{
Other security endpoints: \code{\link{getAllSecurities}},
  \code{\link{getIndicesDataPointHistory}},
  \code{\link{getSecurityById}},
  \code{\link{getSecurityDataPointHistory}},
  \code{\link{getSecurityDataPointNumber}},
  \code{\link{getSecurityDataPointText}},
  \code{\link{getSecurityDividendsLatest}},
  \code{\link{getSecurityIntradayPrices}},
  \code{\link{getSecurityRealtimePrice}},
  \code{\link{getSecurityStockPrices}},
  \code{\link{getSecurityZacksAnalystRatingsSnapshot}}
}
\concept{security endpoints}
