% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.service.R
\name{getSecurityIntradayPrices}
\alias{getSecurityIntradayPrices}
\title{intraday prices for a security over requested time}
\usage{
getSecurityIntradayPrices(identifier, opts = list(),
  max_records = 1000)
}
\arguments{
\item{identifier}{{String} A Security identifier (Ticker, FIGI, ISIN, CUSIP, Intrinio ID)}

\item{opts}{{named list} Optional query params for endpoint}

\item{opts$start_date}{{Date} YYYY-MM-DD Return prices on or after the date}

\item{opts$end_date'{Date}}{YYYY-MM-DD Return prices on or before the date}

\item{opts$start_time}{{Date} YYYY-MM-DD Return prices on or after the date}

\item{opts$end_date':}{{Date} YYYY-MM-DD Return prices on or before the date}

\item{opts$page_size'}{{number = 100} The number of results to return per page}
}
\description{
intraday prices for a security over requested time
}
\seealso{
Other security endpoints: \code{\link{getAllSecurities}},
  \code{\link{getIndicesDataPointHistory}},
  \code{\link{getSecurityById}},
  \code{\link{getSecurityDataPointHistory}},
  \code{\link{getSecurityDataPointNumber}},
  \code{\link{getSecurityDataPointText}},
  \code{\link{getSecurityDividendsLatest}},
  \code{\link{getSecurityEarningsLatest}},
  \code{\link{getSecurityRealtimePrice}},
  \code{\link{getSecurityStockPrices}},
  \code{\link{getSecurityZacksAnalystRatingsSnapshot}}
}
\concept{security endpoints}
