% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security.service.R
\name{getSecurityStockPrices}
\alias{getSecurityStockPrices}
\title{end-of-day prices for a security over requested time}
\usage{
getSecurityStockPrices(identifier, opts = list())
}
\arguments{
\item{identifier}{{String} A Security identifier (Ticker, FIGI, ISIN, CUSIP, Intrinio ID)}

\item{opts}{{named list} Optional query params for endpoint}

\item{opts$start_date}{{Date} YYYY-MM-DD Return prices on or after the date}

\item{opts$end_date':}{{Date} YYYY-MM-DD Return prices on or before the date}

\item{opts$frequency:}{{String = "daily"} daily | weekly | monthly | quarterly | yearly Return stock prices in the given frequency}

\item{opts$page_size':}{{number = 100} The number of results to return per page}
}
\description{
end-of-day prices for a security over requested time
}
\seealso{
Other security endpoints: \code{\link{getAllSecurities}},
  \code{\link{getSecurityById}},
  \code{\link{getSecurityDataPointHistory}},
  \code{\link{getSecurityDataPointNumber}},
  \code{\link{getSecurityDataPointText}},
  \code{\link{getSecurityDividendsLatest}},
  \code{\link{getSecurityEarningsLatest}},
  \code{\link{getSecurityRealtimePrice}},
  \code{\link{getSecurityZacksAnalystRatingsSnapshot}}
}
\concept{security endpoints}
