% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.service.R
\name{getStandardizedFinancialsTable}
\alias{getStandardizedFinancialsTable}
\title{Returns professional-grade standardized historical financial data for the given \code{id}.}
\usage{
getStandardizedFinancialsTable(id)
}
\arguments{
\item{id}{{String}  The Intrinio ID for the Fundamental or lookup code (ticker-statement-year-period)
for the Fundamental}
}
\value{
standardized_financials {tibble} formated as wide table with one column per metric tag
}
\description{
Returns professional-grade historical financial data.
This data is standardized, cleansed and verified to ensure the highest quality
data sourced directly from the XBRL financial statements. The primary purpose
of standardized financials are to facilitate comparability across a single
company’s fundamentals and across all companies' fundamentals.
}
\seealso{
Other fundamental endpoints: \code{\link{getFundamentalById}},
  \code{\link{getReportedFinancialsTable}},
  \code{\link{getReportedFinancials}},
  \code{\link{getStandardizedFinancials}},
  \code{\link{lookupFundamental}}
}
\concept{fundamental endpoints}
