% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrino.service.R
\name{intrino_fetch}
\alias{intrino_fetch}
\title{Perform a get request to an endpoint on the intrino server}
\usage{
intrino_fetch(endpoint, next_page = NULL)
}
\arguments{
\item{endpoint}{a "url" object or a character string which will form the variable part of the endpoint URL
for a discussion of url objects see httr::build_url}
}
\value{
a list, class iex_api, with keys of "status,content,url,response
}
\description{
Perform a get request to an endpoint on the intrino server
}
