% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrinoR.R
\docType{package}
\name{intrinoR}
\alias{intrinoR}
\alias{intrinoR-package}
\title{INTRINO Market Data Api}
\description{
A modern Financial Data API and bulk data marketplace
Intrinio data is delivered in modern formats for today's
developers and analysts, including JSON, Excel, Python, Ruby,
R, Java, CSV, WebSockets and more.
}
\details{
\quote{Intrinio is a financial data platform built for people who
refuse to settle for average data. People who demand better.
People like you.}
}
\seealso{
\url{https://intrinio.com/}
}
\author{
Bruce C. Schardt, \email{schardt.bruce.curtis@gmail.com}
}
\keyword{internal}
