% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.service.R
\name{lookupFundamental}
\alias{lookupFundamental}
\title{Returns the fundamental financial data for the Company with the given `identifier` and with the given parameters}
\usage{
lookupFundamental(identifier, statement_code, fiscal_year, fiscal_period)
}
\arguments{
\item{identifier}{{String}  A Company identifier (Ticker, CIK, LEI, Intrinio ID)}

\item{statement_code}{{String} income_statment | balance_sheet_statement | cash_flow_statement | calculations The statement code}

\item{fiscal_year}{{Number} fiscalYear The fiscal year}

\item{fiscal_period}{{String} Q1TTM | Q2TTM | Q3TTM | FY | Q1 | Q2 | Q3 | Q4 | Q2YTD | Q3YTD The fiscal period}
}
\description{
Returns the fundamental financial data for the Company with the given `identifier` and with the given parameters
}
\seealso{
Other fundamental endpoints: \code{\link{getFundamentalById}},
  \code{\link{getReportedFinancialsTable}},
  \code{\link{getReportedFinancials}},
  \code{\link{getStandardizedFinancialsTable}},
  \code{\link{getStandardizedFinancials}}
}
\concept{fundamental endpoints}
