% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrino.service.R
\name{print.intrino_api}
\alias{print.intrino_api}
\title{print S3 function for intrino_api class}
\usage{
\method{print}{intrino_api}(x, ...)
}
\description{
print S3 function for intrino_api class
}
\keyword{internal}
