% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{useProduction}
\alias{useProduction}
\title{use the intrino production server}
\usage{
useProduction()
}
\description{
When the api calls are directed to the sandbox server there are a limited number of symbols, which return data.
These include the dow30, "AAPL", etc.
}
\details{
Accessing the production data requires a paid subscription
}
\seealso{
\url{https://product.intrinio.com/marketplace/financial-data?category=all&geography=all}
}
